updates to website (#107)
* Corrected errors and updated items

* updating den leaders on den pages

* updated bear page and contact page

* Add Outlaw Derby Page to Website

I think it would be good to have a page on our website for the Outlaw Derby. I created this from scratch. Let me know what you think.

* revised outlaw derby page

I made the suggested changes.

* renamed file to index.md

* Edited paged to update current leaders and other info

We MUST remove that we are meeting online from the home page. I have edited the index.md page in the public folder but it has not been removed.

* Update index.md

---------

Co-authored-by: Thomas Cherry <thomas.cherry@gmail.com>